+++ surfit/src/surfit/cntr.cpp
@@ -138,7 +138,7 @@ cntr_garbage cntr_garb;
 
 #ifndef DOXYGEN_SHOULD_SKIP_THIS
 struct pos_z_w {
-	int pos;
+	size_t pos;
 	REAL z;
 	REAL w;
 };
@@ -150,14 +150,14 @@ bool ptr_pos_z_w_less(const pos_z_w & it1, const pos_z_w & it2)
 	return it1.pos < it2.pos;
 };
 
-bool add_mean_val(d_grid * grd, int x, int y,
+bool add_mean_val(d_grid * grd, size_t x, size_t y,
 		  REAL X1, REAL Y1, REAL Z1,
 		  REAL X2, REAL Y2, REAL Z2,
 		  std::vector<pos_z_w> * data) {
 
 	REAL rx1, rx2;
 	REAL ry1, ry2;
-	int NN = grd->getCountX();
+	size_t NN = grd->getCountX();
 
 	REAL lx1 = X1, lx2 = X2, ly1 = Y1, ly2 = Y2;
 	
@@ -374,9 +374,9 @@ d_points * discretize_cntr(const d_cntr * crv, d_grid * grd, const char * pnts_n
 	pos_z_w elem;
 	REAL total_weight = 0;
 	REAL sum_val = 0;
-	int NN = grd->getCountX();
+	size_t NN = grd->getCountX();
 	
-	for (qq = 0; qq < (int)data->size(); qq++) {
+	for (qq = 0; qq < data->size(); qq++) {
 		elem = (*data)[qq];
 		if (prev_pos == elem.pos) {
 			sum_val += elem.w*elem.z;
@@ -384,8 +384,8 @@ d_points * discretize_cntr(const d_cntr * crv, d_grid * grd, const char * pnts_n
 		} else {
 			// write
 			if (prev_pos != UINT_MAX) {
-				int pos_x = prev_pos % NN;
-				int pos_y = (prev_pos - pos_x)/NN;
+				size_t pos_x = prev_pos % NN;
+				size_t pos_y = (prev_pos - pos_x)/NN;
 				REAL x, y, z;
 				grd->getCoordNode(pos_x, pos_y, x, y);
 				z = sum_val / total_weight;
@@ -400,8 +400,8 @@ d_points * discretize_cntr(const d_cntr * crv, d_grid * grd, const char * pnts_n
 		prev_pos = elem.pos;
 	};
 
-	int pos_x = prev_pos % NN;
-	int pos_y = (prev_pos - pos_x)/NN;
+	size_t pos_x = prev_pos % NN;
+	size_t pos_y = (prev_pos - pos_x)/NN;
 	REAL x, y, z;
 	grd->getCoordNode(pos_x, pos_y, x, y);
 	z = sum_val / total_weight;
